requirements.txt
# Core
torch>=2.0
torchvision>=0.15
transformers>=4.35
accelerate>=0.25

# Computer Vision
opencv-python>=4.8
Pillow>=10.0
albumentations>=1.3

# Color extraction
scikit-learn>=1.3
scikit-image>=0.21
colormath>=3.0

# Brand detection
ultralytics>=8.0  # YOLOv8

# Data & experiment tracking
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
tqdm>=4.65

# Datasets
pycocotools>=2.0  # Para ModaNet (formato COCO)

# Notebooks
jupyter>=1.0
ipywidgets>=8.0

# Testing
pytest>=7.0
pytest-cov>=4.0
